HW3/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cryptography==38.0.2
cycler==0.11.0
flatbuffers==22.9.24
fonttools==4.37.4
frozenlist==1.3.1
gast==0.4.0
google-auth-oauthlib==0.4.6
google-auth==2.13.0
google-pasta==0.2.0
grpcio==1.46.4
h5py==3.6.0
idna==3.4
importlib-metadata==4.11.4
joblib==1.2.0
keras-preprocessing==1.1.2
keras==2.10.0
kiwisolver==1.4.4
libclang==14.0.6
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.6.1
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pillow==9.2.0
pip==22.3
protobuf==3.19.6
pyasn1-modules==0.2.7
pyasn1==0.4.8
pycparser==2.21
pyjwt==2.5.0
pyopenssl==22.1.0
pyparsing==3.0.9
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-nlp==0.5.0
pytz==2022.5
pyu2f==0.1.5
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
setuptools==65.5.0
six==1.16.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
tensorflow==2.9.1
termcolor==2.0.1
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing-extensions==4.4.0
urllib3==1.26.11
werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.9.0